*DONT PUSH TO MASTER AS-IS*
I made the change that psuhes the userinput to the innertables, which is good. Problem is that this let me discover that we have a bug where updates to the userinput don't get passed all the way down. Not sure if in resultTable or innerTable, I think its innerTable. I'm gonna strart looking there, but I don't plan to look too long, I just was trying to get the updated innertables in place.